milestone-2-containerize-the-api/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.1
pytest==8.3.3
Flask-Migrate==4.0.7
mysqlclient==2.2.4
